kernel: move some HPC-GAP functions to region.c
They were in boehm_gc.c, but in reality these functions
are not Bohem specific but rather HPC-GAP specific.
